// src/app/models/landDetail.ts
export interface LandDetail {
    id: number;
    address: string;
    parcelNo: number;
    islandNo: number;
    sheetNo: number;
    floorEquivalent: boolean;
    listingId: number;
    userId: number;
    firstName: string;
    lastName: string;
    userEmail: string;
    title: string;
    description: string;
    cityName: string;
    districtName: string;
    listingTypeName: string;
    price: number;
    date: string;
    squareMeter: number;
}