// src/app/models/updateLand.ts
export interface UpdateLand {
    cityId: number;
    listingTypeId: number;
    propertyTypeId: number;
    districtId: number;
    title: string;
    description: string;
    price: number;
    squareMeter: number;
    status: boolean;
    id: number;
    address: string;
    parcelNo: number;
    islandNo: number;
    sheetNo: number;
    floorEquivalent: boolean;
  }
  